use longs when working with numbers of objects
i think int could be 16-bit in c99, and more than (2**15)-1 = 32767
objects isn't that many... but long is at least 32 bits, and
(2**31)-1 = 2147483647 objects is many

+++ src/atomicincrdecr.h
@@ -2,7 +2,7 @@
 atomic incrementing and decrementing for thread-safe reference counting
 
 example:
-	int var = 0;
+	long var = 0;
 
 	void threadsafe_func(void)
 	{
